MFB-1676: move the CCDF family to cross_white_label/
Pilot for the phase 3 layout: base.py holds the shared class, one file per
white label beside it. Registry keys and class names verified identical to
main before and after.
