+++ .gitignore
@@ -59,3 +59,4 @@ dump.rdb
 # Local research artifacts and one-off reports (belong in Linear/Outline, not git)
 PE_LATENCY_REPORT.md
 *_scoping.csv
+venv
